Move all logging related code from aiida/__init__.py to aiida.common.log
With the new configure_logging function, in the global init
file we simply call this function, which will setup all the
necessary code for logging and configures all the loggers
from the default dictionary
