+++ aiida/__init__.py
@@ -7,10 +7,9 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
-import logging
 import warnings
-from logging import config
-from aiida.common import setup
+from aiida.common.log import configure_logging
+from aiida.common.setup import get_property
 
 __copyright__ = u"Copyright (c), This file is part of the AiiDA platform. For further information please visit http://www.aiida.net/. All rights reserved."
 __license__ = "MIT license, see LICENSE.txt file."
@@ -19,99 +18,10 @@ __authors__ = "The AiiDA team."
 __paper__ = """G. Pizzi, A. Cepellotti, R. Sabatini, N. Marzari, and B. Kozinsky, "AiiDA: automated interactive infrastructure and database for computational science", Comp. Mat. Sci 111, 218-230 (2016); http://dx.doi.org/10.1016/j.commatsci.2015.09.013 - http://www.aiida.net."""
 __paper_short__ = """G. Pizzi et al., Comp. Mat. Sci 111, 218 (2016)."""
 
+# Configure the default logging
+configure_logging()
 
-# Custom logging level, intended specifically for informative log messages
-# reported during WorkChains and Workflows. We want the level between INFO(20)
-# and WARNING(30) such that it will be logged for the default loglevel, however
-# the value 25 is already reserved for SUBWARNING by the multiprocessing module.
-LOG_LEVEL_REPORT = 23
-logging.addLevelName(LOG_LEVEL_REPORT, 'REPORT')
-
-
-# A logging filter that can be used to disable logging
-class NotInTestingFilter(logging.Filter):
-
-    def filter(self, record):
-        from aiida import settings
-        return not settings.TESTING_MODE
-
-# A sample logging configuration. The only tangible logging
-# performed by this configuration is to send an email to
-# the site admins on every HTTP 500 error when DEBUG=False.
-# See http://docs.djangoproject.com/en/dev/topics/logging for
-# more details on how to customize your logging configuration.
-LOGGING = {
-    'version': 1,
-    'disable_existing_loggers': True,
-    'formatters': {
-        'verbose': {
-            'format': '%(levelname)s %(asctime)s %(module)s %(process)d '
-                      '%(thread)d %(message)s',
-        },
-        'halfverbose': {
-            'format': '%(asctime)s, %(name)s: [%(levelname)s] %(message)s',
-            'datefmt': '%m/%d/%Y %I:%M:%S %p',
-        },
-    },
-    'filters': {
-        'testing': {
-            '()': NotInTestingFilter
-        }
-    },
-    'handlers': {
-        'console': {
-            'level': 'DEBUG',
-            'class': 'logging.StreamHandler',
-            'formatter': 'halfverbose',
-            'filters': ['testing']
-        },
-        'daemon_logfile': {
-            'level': 'DEBUG',
-            'formatter': 'halfverbose',
-            'class': 'logging.handlers.RotatingFileHandler',
-            'filename': setup.DAEMON_LOG_FILE,
-            'encoding': 'utf8',
-            'maxBytes': 100000,
-        },
-        'dblogger': {
-            # setup.get_property takes the property from the config json file
-            # The key used in the json, and the default value, are
-            # specified in the _property_table inside aiida.common.setup
-            # NOTE: To modify properties, use the 'verdi devel setproperty'
-            #   command and similar ones (getproperty, describeproperties, ...)
-            'level': setup.get_property('logging.db_loglevel'),
-            'class': 'aiida.utils.logger.DBLogHandler',
-        },
-    },
-    'loggers': {
-        'aiida': {
-            'handlers': ['console', 'dblogger'],
-            'level': setup.get_property('logging.aiida_loglevel'),
-            'propagate': False,
-        },
-        'paramiko': {
-            'handlers': ['console'],
-            'level': setup.get_property('logging.paramiko_loglevel'),
-            'propagate': False,
-        },
-        'alembic': {
-            'handlers': ['console'],
-            'level': setup.get_property('logging.alembic_loglevel'),
-            'propagate': False,
-        },
-        'sqlalchemy': {
-            'handlers': ['console'],
-            'level': setup.get_property('logging.sqlalchemy_loglevel'),
-            'propagate': False,
-            'qualname': 'sqlalchemy.engine',
-        },
-    },
-}
-
-# Configure the global logger through the LOGGING dictionary
-logging.config.dictConfig(LOGGING)
-
-if setup.get_property("warnings.showdeprecations"):
+if get_property("warnings.showdeprecations"):
     # print out the warnings coming from deprecation
     # in Python 2.7 it is suppressed by default
     warnings.simplefilter('default', DeprecationWarning)
